Add Cloud component and adjust Sun position
Added animated cloud component as inline SVG. The cloud moves from
one side of the screen to the other. The size, position, and even color of the cloud
can be manipulated.

Also set the Sun and Cloud components to position: fixed

Sun animation now moves according to viewport height, rather than width,
and set sun margin-top to 100vh to keep it on bottom of screen